Preserve configured private upstream support
